Extract shared `ab-aligned-buffer` crate out of `ab-executor-slots`
