+++ crates/executor/ab-executor-native/src/context.rs
@@ -4,7 +4,7 @@ use crate::context::ffi_call::make_ffi_call;
 use ab_contracts_common::env::{EnvState, ExecutorContext, MethodContext, PreparedMethod};
 use ab_contracts_common::method::{ExternalArgs, MethodFingerprint};
 use ab_contracts_common::{Address, ContractError, ExitCode, ShardIndex};
-use ab_executor_slots::slots::NestedSlots;
+use ab_executor_slots::NestedSlots;
 use ab_system_contract_address_allocator::ffi::allocate_address::AddressAllocatorAllocateAddressArgs;
 use halfbrown::HashMap;
 use std::cell::UnsafeCell;
