+++ crates/shared/README.md
@@ -1,0 +1,3 @@
+# Shared
+
+This part of the repository shared crates used across different parts of the codebase
