+++ crates/shared/ab-aligned-buffer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ab-executor-slots"
+name = "ab-aligned-buffer"
 description = ""
 license = "0BSD"
 version = "0.0.1"
@@ -14,11 +14,7 @@ include = [
 all-features = true
 
 [dependencies]
-ab-aligned-buffer = { workspace = true }
-ab-contracts-common = { workspace = true, features = ["executor"] }
-smallvec = { workspace = true }
-take_mut = { workspace = true }
-tracing = { workspace = true }
+ab-contracts-io-type = { workspace = true }
 
 [lints]
 workspace = true
