+++ crates/shared/ab-aligned-buffer/src/tests.rs
@@ -1,4 +1,4 @@
-use crate::aligned_buffer::{AlignedBytes, OwnedAlignedBuffer, SharedAlignedBuffer};
+use crate::{AlignedBytes, OwnedAlignedBuffer, SharedAlignedBuffer};
 use alloc::vec;
 
 const EXPECTED_ALIGNMENT: usize = size_of::<AlignedBytes>();
